Add a way to run commands from the HE page
Move deployment to a dialog, and the HE page will just display
basic information by default. Add dialogs for getting the VM
status and setting maintenance level.

Interim patch. Resolves the RFE, but followup to move logic out
of the page into the model as much as possible, like the rest of
the codebase

Change-Id: I5a7ddf9b7959e861a64244b1c6cd3b449955f63c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1208780
Signed-off-by: rbarry@redhat.com <rbarry@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40897
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>