vm.conf doesn't exist on 3.6, read from hosted-engine.conf
Since vm.conf doesn't exist, the TUI will always show "Not
configured", even if it is. Pull the fqdn out of
hosted-engine.conf and use that.

We could query VDSM directly, but we run the risk of the VM not
being started or not running on this host and having the same
problem.

Change-Id: Ie280a6efa2bf98860b66d2e82c5ddf32d09aa8e7
Signed-off-by: Ryan Barry <rbarry@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47675
Continuous-Integration: Jenkins CI
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>